🧬 Move the common function for pre-process by feature spec to global level test utility module.
