+++ test/integration_test/server/init.py
@@ -10,7 +10,7 @@ import pytest
 import pymock_api.server as mock_server
 
 from ..._file_utils import MockAPI_Config_Yaml_Path, yaml_factory
-from .._spec import run_test
+from ..._spec import run_test
 
 LOAD_APP_TYPE = Type[mock_server.load_app]
